requirements.txt
flask==3.0.2
waitress==3.0.0
torch==2.2.0
torchvision==0.17.0
numpy<2.0.0
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
python-dotenv>=0.19.0
requests>=2.26.0
werkzeug>=2.0.0
matplotlib>=3.2.2
seaborn>=0.11.0
pandas>=1.1.4
tqdm>=4.64.0
pymongo>=4.0.0